refactor: move interfaces and helper functions to respective files
